refactor(014): extract audit-barrage-fleet.ts — line-cap relief
audit-barrage.ts grew to 616 lines across US1 + AUDIT-20260611-03;
the fleet-degradation/floor logic (zero-output classification,
FleetFloorEvaluation, renderFleetWarnings, deriveBarrageExitCode)
moves verbatim to audit-barrage-fleet.ts (225 lines) with re-exports
preserving every import path. audit-barrage.ts is 417 lines. Zero
behavior change; barrage suites green (36/36) with no test edits;
tsc clean.
